Lonely_ViT_VS/requirements.txt
# ViT-VS Standalone Requirements
# Complete dependencies for ViT Visual Servoing System

# ========================================
# CORE PYTORCH ECOSYSTEM (DINOv2 & ViT)
# ========================================
# PyTorch con supporto CUDA per accelerazione GPU
--index-url https://download.pytorch.org/whl/cu118
torch>=2.7.0+cu118
torchvision>=0.22.0+cu118

# ========================================
# COMPUTER VISION & IMAGE PROCESSING
# ========================================
opencv-python>=4.8.0
pillow>=10.0.0

# ========================================
# SCIENTIFIC COMPUTING
# ========================================
numpy>=1.24.0
scipy>=1.10.0

# ========================================
# VISUALIZATION & PLOTTING
# ========================================
matplotlib>=3.7.0

# ========================================
# CONFIGURATION & UTILITIES
# ========================================
pyyaml>=6.0.0

# ========================================
# WEB SERVER & API
# ========================================
flask>=2.3.0
requests>=2.31.0

# ========================================
# DEEP LEARNING MODEL UTILITIES
# ========================================
# TIMM for alternative ViT models
timm>=0.9.0

# Note: DINOv2 viene caricato tramite torch.hub

# ========================================
# SYSTEM REQUIREMENTS
# ========================================
# GPU: NVIDIA GPU with CUDA support (recommended)
# VRAM: 20GB+
# Python: 3.12+
# OS: Windows/Linux/macOS
